llama_api/requirements.txt
flask==2.3.3
llama-cpp-python==0.2.11
requests==2.31.0
